# src/controllers/resturant.controller.ts
import { IResturant } from "../types/documents/resturant.document";
import { ResturantRepo } from "../repositories/resturant.repositories";
import { Get, Route, Tags, Post, Body, Path, Put, Delete, SuccessResponse, Security } from "tsoa";
import CustomError from '../utils/error';
import { IRegisterResturantReq, IDeleteResturantReq, IGetResturantReq, ICheckNearbyResturantReq } from "../types/requests/resturant.request";
import { ICheckNearbyResturantRes, IRegisterResturantRes } from "../types/responses/resturant.response";
const geolib = require('geolib')
@Route('resturant')
@Tags('resturant')

export class ResturantController {

    @Security('api_key')
    @Post('registerResturant')
    async registerResturant(@Body() regReq: IRegisterResturantReq): Promise<IRegisterResturantRes> {
        const newResturant: IResturant = await new ResturantRepo().registerResturant(<IResturant>(regReq));
        return <IRegisterResturantRes> (newResturant)
    }

    @Security('api_key')
    @Delete('deleteResturant')
    @SuccessResponse('200', 'Resturant Deleted')
    async deleteResturant(@Body() delReq: IDeleteResturantReq) {
        return await new ResturantRepo().deleteResturant(delReq._id)
    }

    @Post('getResturant')
    async getResturant(@Body() getReq: IGetResturantReq): Promise<IRegisterResturantRes> {
        const newResturant: IResturant = <any> await new ResturantRepo().getResturant(getReq._id);
        if(newResturant === null) throw new CustomError(404, 'Resturant not found');
        return <IRegisterResturantRes> newResturant
    }

    @Post('checkNearByResturants')
    async checkNearByResturants(@Body() location: ICheckNearbyResturantReq): Promise<ICheckNearbyResturantRes[]> {
        const resturants: IResturant[] = await new ResturantRepo().checkNearByResturants();
        let nearResturants: any [] = [];
        resturants.map(resturant => {
            const distance = geolib.getDistance({ latitude: resturant.location.latitude, longitude: resturant.location.longitude },
                { latitude: location.location.latitude, longitude: location.location.longitude })
                if (distance < location.radius) {
                    const nearResturant = {
                        name: resturant.name,
                        distance: distance,
                        location: {
                            latitude: resturant.location.latitude,
                            longitude: resturant.location.longitude
                        }
                    }
                    nearResturants.push(nearResturant)
                }
               
        })
        return <ICheckNearbyResturantRes[]>nearResturants;
    }

}