# src/types/documents/user.document.ts
import { Document } from 'mongoose';

export interface IUser extends Document {
    _id: string;
    name: string;
    email: string;
    password: string;
    createdAt?: string;
    updatedAt?: string;
}